+++ src/main/java/ru/novemis/rpgapp/dto/announcement/CommentRqDto.kt
@@ -1,7 +1,7 @@
 package ru.novemis.rpgapp.dto.announcement
 
 data class CommentRqDto(
-        var authorId: Int = 1,
+        var authorId: Long = 1,
         var announcementId: String = "",
         var text: String = ""
 )
